+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-# Unreleased
+# 5.0.0
 
-* Fix: event.origin and event.environment tags have wrong value for iOS (#365)
+* Sound null safety
+* Fix: event.origin and event.environment tags have wrong value for iOS (#365) and (#369)
 * Fix: Fix deprecated `registrar.messenger` call in `SentryFlutterWeb` (#364)
 * Fix: Enable breadcrumb recording mechanism based on platform (#366)
 * Feat: Send default PII options (#360)
 * Bump: sentry-cocoa to v6.2.1 (#360)
 * Feat: Migration from `package_info` to `package_info_plus` plugin (#370)
-* Fix: missing event.origin and event.environment tags for non-native events on iOS (#369)
 * Fix: Set `SentryOptions.debug` in `sentry` (#376)
 * Fix: Read all environment variables in `sentry` (#375)